+++ src/components/Accordion/Accordion.stories.tsx
@@ -138,7 +138,6 @@ const meta: Meta<typeof Accordion> = {
         faqData: faqContentData,
       },
     },
-    ariaLabel: { control: false },
     id: { control: false },
     isDefaultOpen: argsBooleanType(),
     isAlwaysRendered: argsBooleanType(),
@@ -156,7 +155,6 @@ type Story = StoryObj<typeof Accordion>;
 export const WithControls: Story = {
   args: {
     accordionData,
-    ariaLabel: undefined,
     id: "accordion-id",
     isDefaultOpen: false,
     isAlwaysRendered: false,