+++ src/components/MultiSelect/MultiSelect.tsx
@@ -459,7 +459,7 @@ export const MultiSelect: ChakraComponent<
                 panel: accordionPanel,
               },
             ]}
-            ariaLabel={ariaLabelValue}
+            aria-label={ariaLabelValue}
             id={`multi-select-accordion-${id}`}
             isDefaultOpen={isDefaultOpen}
             isAlwaysRendered