+++ src/components/Button/buttonChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2025-05-22",
+    version: "3.6.2",
+    type: "Bug Fix",
+    affects: ["Styles"],
+    notes: ["Fixes the style transitions."],
+  },
+  {
     date: "2025-04-10",
     version: "3.6.0",
     type: "Update",