+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the Storybook font styles.
 - Updates the docs for the `Template` component.
 - Replaces the error thrown with a console.warn for the `Breadcrumbs`, `Heading`, and `Image` components.
+- Updates the grid layout of the content for the `Hero` `"primary"` variant.
 
 ### Removals