+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
   - `StatusBadge`'s "low", "medium", "high" values from `statusBadgeTypeArray`
   - `StatusBadge`'s `level` prop
   - `StyledList`'s "tag", "mini" values in `textSizesArray`
+- Removes explicit `className` and `children` props in favor of expanded prop type definitions
 
 ## Prerelease