+++ CHANGELOG.md
@@ -8,8 +8,9 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
-## Adds
+### Adds
 
+- Adds "ARIA Landmarks" accessibility guide
 - Adds Progressive Enhancement accessibility guide Storybook page.
 
 ### Updates
@@ -24,6 +25,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Removes
 
 - Removes use of `useNYPLBreakpoints` hook where possible without causing breaking change.
+
 
 ## 3.6.1 (April 24, 2025)